requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.1
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
openai>=1.0.0
